refactor /plugins/sortFriendsByDate into /plugins/sortFriendRequests
