+++ src/plugins/sortFriendRequests/README.md
@@ -1,0 +1,3 @@
+# SortFriendRequests
+
+Sorts friend requests by date of receipt, and sorts the friends list by date added
