+++ src/plugins/sortFriendRequests/index.tsx
@@ -1,20 +1,8 @@
 /*
- * Vencord, a modification for Discord's desktop app
- * Copyright (c) 2022 Vendicated and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
+ * Vencord, a Discord client mod
+ * Copyright (c) 2026 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
 
 import "./styles.css";
 
@@ -25,7 +13,7 @@ import { Devs } from "@utils/constants";
 import { classNameFactory } from "@utils/css";
 import definePlugin, { OptionType } from "@utils/types";
 import { User } from "@vencord/discord-types";
-import { DateUtils, RelationshipStore, Text } from "@webpack/common";
+import { Constants, DateUtils, FriendsStore, RelationshipStore, Text } from "@webpack/common";
 import { PropsWithChildren } from "react";
 
 const formatter = new Intl.DateTimeFormat(undefined, {
@@ -46,35 +34,115 @@ const settings = definePluginSettings({
         description: "Show dates on friend requests",
         default: false,
         restartNeeded: true
+    },
+    sortFriendsBy: {
+        type: OptionType.SELECT,
+        description: "Sort order for the friends list (not requests)",
+        default: "default",
+        options: [
+            { label: "Oldest First", value: "oldest" },
+            { label: "Newest First", value: "newest" },
+            { label: "Default Sorting", value: "default" }
+        ]
     }
 });
 
 export default definePlugin({
     name: "SortFriendRequests",
-    authors: [Devs.Megu],
-    description: "Sorts friend requests by date of receipt",
+    authors: [Devs.Megu, Devs["0bi0"]],
+    description: "Sorts friend requests by date of receipt, and sorts the friends list by date added",
     tags: ["Friends", "Organisation"],
     settings,
 
-    patches: [{
-        find: "getRelationshipCounts(){",
-        replacement: {
-            match: /\}\)\.sortBy\((.+?)\)\.value\(\)/,
-            replace: "}).sortBy(row => $self.wrapSort(($1), row)).value()"
+    patches: [
+        {
+            find: "getRelationshipCounts(){",
+            replacement: {
+                match: /\}\)\.sortBy\((.+?)\)\.value\(\)/,
+                replace: "}).sortBy(row => $self.wrapSort(($1), row)).value()"
+            }
+        },
+        {
+            find: "#{intl::FRIEND_REQUEST_CANCEL}",
+            replacement: {
+                predicate: () => settings.store.showDates,
+                match: /(?<=children:\[)\(0,.{0,100}user:\i,hovered:\i.+?(?=,\(0)(?<=user:(\i).+?)/,
+                replace: (children, user) => `$self.WrapperDateComponent({user:${user},children:${children}})`
+            }
+        },
+        {
+            find: "#{intl::FRIENDS_SECTION_ONLINE}),className:",
+            replacement: {
+                match: /,{id:(\i\.\i)\.PENDING,show:.+?className:(\i\.\i)(?=\},\{id:)/,
+                replace: (rest, relationShipTypes, className) => `,{id:${relationShipTypes}.SORT,show:true,className:${className},content:$self.getSortTitle()}${rest}`
+            }
+        },
+        {
+            find: "FRIENDS_SET_SECTION:",
+            replacement: {
+                match: /FRIENDS_SET_SECTION:function\((\i)\){(\i)=\1\.section,(\i)\(\)}/,
+                replace: 'FRIENDS_SET_SECTION:function($1){if($1.section==="SORT"){return $self.cycleSortFriendsBy()}$2=$1.section,$3()}'
+            }
         }
-    }, {
-        find: "#{intl::FRIEND_REQUEST_CANCEL}",
-        replacement: {
-            predicate: () => settings.store.showDates,
-            match: /(?<=children:\[)\(0,.{0,100}user:\i,hovered:\i.+?(?=,\(0)(?<=user:(\i).+?)/,
-            replace: (children, user) => `$self.WrapperDateComponent({user:${user},children:${children}})`
+    ],
+
+    sinceCache: new Map<string, number>(),
+
+    getCachedSince(userId: string): number {
+        let cached = this.sinceCache.get(userId);
+        if (cached === undefined) {
+            cached = getSince({ id: userId } as User).getTime();
+            this.sinceCache.set(userId, cached);
         }
-    }],
+        return cached;
+    },
 
     wrapSort(comparator: Function, row: any) {
-        return row.type === 3 || row.type === 4
-            ? -getSince(row.user)
-            : comparator(row);
+        if (row.type === 3 || row.type === 4) {
+            return -getSince(row.user);
+        }
+
+        if (row.type === 1 && settings.store.sortFriendsBy !== "default") {
+            const state = FriendsStore.getState();
+            const isSortableTab =
+                state.section === "ALL" || state.section === Constants.FriendsSections.ALL ||
+                state.section === "ONLINE" || state.section === Constants.FriendsSections.ONLINE;
+
+            if (isSortableTab) {
+                const timestamp = this.getCachedSince(row.userId);
+                return settings.store.sortFriendsBy === "newest" ? -timestamp : timestamp;
+            }
+        }
+
+        return comparator(row);
+    },
+
+    getSortTitle() {
+        switch (settings.store.sortFriendsBy) {
+            case "newest":
+                return "Sort (newest)";
+            case "oldest":
+                return "Sort (oldest)";
+            case "default":
+            default:
+                return "Sort (default)";
+        }
+    },
+
+    cycleSortFriendsBy() {
+        switch (settings.store.sortFriendsBy) {
+            case "default":
+                settings.store.sortFriendsBy = "oldest";
+                break;
+            case "oldest":
+                settings.store.sortFriendsBy = "newest";
+                break;
+            case "newest":
+            default:
+                settings.store.sortFriendsBy = "default";
+                break;
+        }
+        FriendsStore.emitChange();
     },
 
     WrapperDateComponent: ErrorBoundary.wrap(({ user, children }: PropsWithChildren<{ user: User; }>) => {
@@ -88,5 +156,17 @@ export default definePlugin({
                 </TooltipContainer>
             )}
         </div>;
-    }, { noop: true })
+    }, { noop: true }),
+
+    start() {
+        Constants.FriendsSections.SORT = "SORT";
+        try {
+            const friendIds = RelationshipStore.getFriendIDs();
+            for (const id of friendIds) {
+                this.getCachedSince(id);
+            }
+        } catch (e) {
+            console.error("[SortFriendRequests] Failed to prime friend-since cache:", e);
+        }
+    }
 });
