--- src/plugins/sortFriendsByDate/README.md
@@ -1,5 +1,0 @@
-# SortFriendsByDate
-
-Sorts friends by oldest/newest/default in the friends list
-
-![](https://github.com/user-attachments/assets/0a683c6d-521d-405c-aee5-43e764f11ed0)
